platform: generic: remove ligh_ext.c
There is a lot of code in light_ext.c that is duplicated in
thead-generic.c, so delete it.
