+++ platform/generic/thead/objects.mk
@@ -13,4 +13,3 @@ platform-objs-$(CONFIG_THEAD_C9XX_ERRATA) += thead/thead_c9xx_errata_tlb_flush.o
 carray-platform_override_modules-$(CONFIG_PLATFORM_THEAD) += thead_generic
 platform-objs-$(CONFIG_PLATFORM_THEAD) += thead/thead-generic.o
 platform-objs-$(CONFIG_PLATFORM_THEAD) += thead/light_c910.o
-platform-objs-$(CONFIG_PLATFORM_THEAD) += thead/light_ext.o
