+++ platform/generic/thead/thead-generic.c
@@ -10,16 +10,36 @@
 #include <thead/c9xx_encoding.h>
 #include <thead/c9xx_errata.h>
 #include <thead/c9xx_pmu.h>
+#include <thead/light/asm.h>
+#include <sbi/riscv_io.h>
 #include <sbi/sbi_const.h>
 #include <sbi/sbi_console.h>
+#include <sbi/sbi_ecall.h>
 #include <sbi/sbi_platform.h>
 #include <sbi/sbi_scratch.h>
 #include <sbi/sbi_string.h>
 #include <sbi/sbi_hsm.h>
 #include <sbi_utils/fdt/fdt_helper.h>
 
+#define SBI_EXT_VENDOR_SMC      (SBI_EXT_VENDOR_START + 0)
+#define SBI_EXT_VENDOR_PMU      (SBI_EXT_VENDOR_START + 1)
+#define SBI_EXT_VENDOR_PMP      (SBI_EXT_VENDOR_START + 2)
+
+#define CSR_MCOUNTERWEN  0x7c9
+#define PMP_BASE_ADDR			0xffdc020000UL
+#define PMP_SIZE_PER_CORE		0x4000UL
+#define TCM0_START_ADDR			0xffe0180000UL
+#define TCM0_END_ADDR			0xffe01c0000UL
+#define TCM1_START_ADDR			0xffe01c0000UL
+#define TCM1_END_ADDR			0xffe0200000UL
+#define RESERVED_START_ADDR		0xffe0200000UL
+#define RESERVED_END_ADDR		0xffe1000000UL
+#define PMP_ENTRY_BASE_ADDR		0x100UL
+#define PMP_ENTRY_START_ADDR(n)	(PMP_BASE_ADDR + PMP_ENTRY_BASE_ADDR + (n * 8))
+#define PMP_ENTRY_END_ADDR(n)	(PMP_ENTRY_START_ADDR(n) + 4)
+#define PMP_ENTRY_CFG_ADDR(n)	(PMP_BASE_ADDR + ((n / 4) * 4))
+
 extern const struct sbi_hsm_device light_ppu;
-extern struct sbi_ecall_extension ecall_light;
 static bool has_ecall_light;
 struct thead_generic_quirks {
 	u64	errata;
@@ -186,6 +206,126 @@ static void sbi_thead_pmu_set(unsigned long type, unsigned long idx, unsigned lo
 		break;
 	}
 }
+
+static void sbi_thead_reserved_pmp_set(void)
+{
+	unsigned int num, reg_val;
+
+	for (num = 0; num < 4; num++) {
+		/*	pmp entry 28 for reserved memory	*/
+		writel(RESERVED_START_ADDR >> 12, (void *)(PMP_ENTRY_START_ADDR(28) + num*PMP_SIZE_PER_CORE));
+		writel(RESERVED_END_ADDR >> 12, (void *)(PMP_ENTRY_END_ADDR(28) + num*PMP_SIZE_PER_CORE));
+
+		/*	pmp entry 28 config	*/
+		reg_val = readl((void *)(PMP_ENTRY_CFG_ADDR(28) + num*PMP_SIZE_PER_CORE));
+		reg_val = (reg_val & 0xffffff00) | 0x040;
+		writel(reg_val, (void *)((PMP_ENTRY_CFG_ADDR(28) + num*PMP_SIZE_PER_CORE)));
+	}
+
+	sync_is();
+}
+
+static void sbi_thead_tcm0_pmp_set(unsigned long auth)
+{
+	sbi_printf("%s: auth:%lx \n", __func__, auth);
+	unsigned int num, reg_val;
+
+	reg_val = readl((void *)PMP_ENTRY_START_ADDR(26));
+
+	if (reg_val != TCM0_START_ADDR >> 12)
+		for(num = 0; num < 4; num++) {
+			/*	pmp entry 26 for dsp tcm0	*/
+			writel(TCM0_START_ADDR >> 12, (void *)(PMP_ENTRY_START_ADDR(26) + num*PMP_SIZE_PER_CORE));
+			writel(TCM0_END_ADDR >> 12, (void *)(PMP_ENTRY_END_ADDR(26) + num*PMP_SIZE_PER_CORE));
+		}
+
+	for(num = 0; num < 4; num++) {
+		/*	pmp entry 26 config	*/
+		reg_val = readl((void *)(PMP_ENTRY_CFG_ADDR(26) + num*PMP_SIZE_PER_CORE));
+		reg_val = (reg_val & 0xff00ffff) | (auth << 16);
+		writel(reg_val, (void *)(PMP_ENTRY_CFG_ADDR(26) + num*PMP_SIZE_PER_CORE));
+	}
+
+	sync_is();
+}
+
+static void sbi_thead_tcm1_pmp_set(unsigned long auth)
+{
+	sbi_printf("%s: auth:%lx \n", __func__, auth);
+	unsigned int num, reg_val;
+
+	reg_val = readl((void *)PMP_ENTRY_START_ADDR(27));
+	if (reg_val != TCM1_START_ADDR >> 12)
+		for (num = 0; num < 4; num++) {
+			/*	pmp entry 27 for dsp tcm1	*/
+			writel(TCM1_START_ADDR >> 12, (void *)(PMP_ENTRY_START_ADDR(27) + num*PMP_SIZE_PER_CORE));
+			writel(TCM1_END_ADDR >> 12, (void *)(PMP_ENTRY_END_ADDR(27) + num*PMP_SIZE_PER_CORE));
+		}
+
+	for (num = 0; num < 4; num++) {
+		/*	pmp entry 27 config	*/
+		reg_val = readl((void *)(PMP_ENTRY_CFG_ADDR(27) + num*PMP_SIZE_PER_CORE));
+		reg_val = (reg_val & 0x00ffffff) | (auth << 24);
+		writel(reg_val, (void *)(PMP_ENTRY_CFG_ADDR(27) + num*PMP_SIZE_PER_CORE));
+	}
+
+	sync_is();
+}
+
+static void sbi_thead_pmp_set(unsigned long idx, unsigned long auth)
+{
+	unsigned int reg_val;
+
+	if (idx !=0 && idx != 1)
+		return;
+
+	/*	read pmp entry 28	*/
+	reg_val = readl((void *)PMP_ENTRY_START_ADDR(28));
+
+	if (reg_val != RESERVED_START_ADDR >> 12)
+		sbi_thead_reserved_pmp_set();
+
+	switch (idx) {
+	case 0:
+		sbi_thead_tcm0_pmp_set(auth);
+		break;
+	case 1:
+		sbi_thead_tcm1_pmp_set(auth);
+		break;
+	default:
+		break;
+	}
+}
+
+static int sbi_ecall_light_handler(unsigned long extid, unsigned long funcid,
+		       struct sbi_trap_regs *regs,
+		       struct sbi_ecall_return *out)
+{
+	switch(extid) {
+	case SBI_EXT_VENDOR_PMU:
+		sbi_thead_pmu_set(regs->a0, regs->a1, regs->a2);
+		break;
+	case SBI_EXT_VENDOR_PMP:
+		sbi_thead_pmp_set(funcid, regs->a0);
+		break;
+	}
+	return 0;
+}
+
+struct sbi_ecall_extension ecall_light;
+
+static int sbi_ecall_light_register_extensions(void)
+{
+	return sbi_ecall_register_extension(&ecall_light);
+}
+
+struct sbi_ecall_extension ecall_light = {
+	.name			= "light",
+	.extid_start		= SBI_EXT_VENDOR_PMU,
+	.extid_end		= SBI_EXT_VENDOR_PMP,
+	.register_extensions	= sbi_ecall_light_register_extensions,
+	.handle			= sbi_ecall_light_handler,
+};
 
 static int thead_vendor_ext_provider(long funcid,
 				   struct sbi_trap_regs *regs,
